assembler is done working on basic working on the EMU starting to make an OS working on some basic Input and output (more in FuncCalls.txt) added syntax using vim added more instr (PUSHR POPR ROL ROR JER) the addr bus is now 20 bit long

+++ emu/ConsolePort.cs
@@ -10,8 +10,8 @@ namespace emu
     {
         ushort[] VRam = new ushort[0x6C00];
         public MEM MEM { get; set; }
-        public ushort DataBus { get; set; }
-        public ushort AddrBus { get; set; }
+        public uint DataBus { get; set; }
+        public uint AddrBus { get; set; }
         public bool Read { get; set; }
         public ushort MemBank { get; set; }
