+++ emu/Screen.cs
@@ -9,8 +9,8 @@ namespace emu
     public class Screen : IExpansionPort
     {
         public MEM MEM { get; set; }
-        public ushort DataBus { get; set; }
-        public ushort AddrBus { get; set; }
+        public uint DataBus { get; set; }
+        public uint AddrBus { get; set; }
         public bool Read { get; set; }
         public ushort MemBank { get; set; }
         public void TICK()
